requirements.txt
# ── Server ────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
ollama>=0.2.0
python-dotenv>=1.0.0
httpx>=0.27.0

# ── Search ────────────────────────────────────────────────────────────────────
duckduckgo-search>=6.1.0

# ── Screen Capture ────────────────────────────────────────────────────────────
mss>=9.0.1
Pillow>=10.3.0
pyautogui>=0.9.54

# ── Voice Pipeline ────────────────────────────────────────────────────────────
openwakeword>=0.6.0
faster-whisper>=1.0.2
piper-tts>=1.2.0
sounddevice>=0.4.6
soundfile>=0.12.1
pyaudio>=0.2.14
numpy>=1.26.0
edge-tts>=6.1.9

# ── UI ────────────────────────────────────────────────────────────────────────
customtkinter>=5.2.2
pystray>=0.19.5
pygame>=2.5.2

# ── System Control ────────────────────────────────────────────────────────────
pyperclip>=1.8.2
psutil>=5.9.8
opencv-python>=4.9.0
pywin32>=306

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
